# util/deephash: use sha256x